realestate/requirements.txt
Django==2.2.10
django-extensions==3.1.3
djangorestframework==3.12.2
graphviz==0.16
mysql-connector==2.2.9
Pillow==8.1.2
pydot==1.4.2
pydotplus==2.0.2
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
